+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+#### Dependency Upgrade
+- @twreporter/react-components@^7.1.1
+
+#### Rewrite components with scss with styled-components
+- Rewrite `AuthorsList`, `AuthorPage`, and `TopicLandingPage`
+
 ## Release
 
 ### 4.3.11